fixing token location ranges

+++ alpha-core/common/source_location.cpp
@@ -18,8 +18,8 @@ namespace basecode::common {
 
     source_location::source_location(
             uint32_t line,
-            uint16_t start_column,
-            uint16_t end_column) : _line(line),
+            uint32_t start_column,
+            uint32_t end_column) : _line(line),
                                    _end_column(end_column),
                                    _start_column(start_column) {
     }
@@ -32,19 +32,19 @@ namespace basecode::common {
         _line = value;
     }
 
-    uint16_t source_location::end_column() const {
+    uint32_t source_location::end_column() const {
         return _end_column;
     }
 
-    uint16_t source_location::start_column() const {
+    uint32_t source_location::start_column() const {
         return _start_column;
     }
 
-    void source_location::end_column(uint16_t value) {
+    void source_location::end_column(uint32_t value) {
         _end_column = value;
     }
 
-    void source_location::start_column(uint16_t value) {
+    void source_location::start_column(uint32_t value) {
         _start_column = value;
     }
