+++ alpha-core/common/source_location.h
@@ -21,25 +21,25 @@ namespace basecode::common {
 
         source_location(
             uint32_t line,
-            uint16_t start_column,
-            uint16_t end_column);
+            uint32_t start_column,
+            uint32_t end_column);
 
         uint32_t line() const;
 
         void line(uint32_t value);
 
-        uint16_t end_column() const;
+        uint32_t end_column() const;
 
-        uint16_t start_column() const;
+        uint32_t start_column() const;
 
-        void end_column(uint16_t value);
+        void end_column(uint32_t value);
 
-        void start_column(uint16_t value);
+        void start_column(uint32_t value);
 
     private:
         uint32_t _line = 0;
-        uint16_t _end_column = 0;
-        uint16_t _start_column = 0;
+        uint32_t _end_column = 0;
+        uint32_t _start_column = 0;
     };
 
 };
